♻️  report screen refactoring
