+++ src/views/report/index.vue
@@ -9,7 +9,7 @@
       </div>
     </div>
     <div v-else>
-      <h1>Create your report.</h1>
+      <h1>Enter report details.</h1>
       <div class="container report-page__form" ref="form">
         <div class="company-card">
           <span>TO:</span>
@@ -22,13 +22,14 @@
           <strong> {{ selectedCompany.companyName }} </strong>
         </div>
         <div
-          v-for="element in formElements"
+          v-for="element in formFields"
           class="report-page__element"
           :key="element.id"
         >
           <input-generator
             :element="element"
             :data-field="element.id"
+            @onUpdate="onInputChange"
             v-model="element.value"
           />
           <div
@@ -44,6 +45,7 @@
                 v-if="subfield.parentField === element.value"
                 :element="subfield"
                 :data-field="subfield.id"
+                @onUpdate="onInputChange"
                 v-model="subfield.value"
               />
             </div>
@@ -55,35 +57,11 @@
           @click="goLastStep"
         >Go Last Step</button>
       </div>
-      <div
-        class="report-page__warning"
-        v-if="passcode"
-      >
-        <img src="@/assets/warning.svg" alt="warning">
-        <div class="report-page__content">
-          <strong>Save Anonymous Login Info</strong>
-          <p><b>Be careful</b>! You are creating this case with an anonymous profile. At the bottom, you see a passcode. You will use these passcode to login into the app, which allows you to follow your case's further updates.</p>
-          <div class="report-page__passcode">
-            <span>PASSCODE</span> <input ref="passcode" :value="passcode" readonly>
-            <button @click="copyToClipboard">COPY</button>
-          </div>
-          <div class="checkbox checkbox--signal">
-            <input type="checkbox" v-model="conditionsAccepted">
-            <span>
-              I understand that if I forget or lose the passcode, I have to create my case report from the beginning as the system ensures my safety, security and anonymousness!
-            </span>
-          </div>
-          <button
-            class="button button--red"
-            :disabled="!conditionsAccepted"
-            :class="{
-              'button--signal': conditionsAccepted,
-              'button--progress': loading
-            }"
-            @click="finishCaseCreation"
-          >Finish and Go Case Detail</button>
-        </div>
-      </div>
+      <warning
+        :selected-company="selectedCompany"
+        :loading="loading"
+        :onFinish="finishCaseCreation"
+      />
     </div>
   </div>
 </template>
@@ -93,9 +71,11 @@
  * !TODO(1): The form should be resetted once the user update the FeedbackType. (to avoid some bugs)
  * !TODO(2): Required messages shouldn't be visible at the first load
  */
+import mapValues from 'lodash/mapValues';
 import InputGenerator from '@/components/InputGenerator.vue';
 import CompanySelect from '@/components/CompanySelect/index.vue';
-import formElements from './form/index';
+import FORM_FIELDS from './codables/index';
+import Warning from './components/Warning.vue';
 import API from '../../api';
 import { firebase } from '../../firebase';
 
@@ -104,20 +84,20 @@ export default {
   components: {
     InputGenerator,
     CompanySelect,
+    Warning,
   },
   data() {
     return {
       loading: false,
-      formElements,
+      formFields: FORM_FIELDS,
       selectedCompany: null,
       enteredData: [],
-      conditionsAccepted: false,
       passcode: null,
     };
   },
   computed: {
     isFormValid() {
-      const hasFormElementsValid = Object.values(this.formElements).map((field) => {
+      const hasFormElementsValid = Object.values(this.formFields).map((field) => {
         if (field.subfields) {
           Object.values(field.subfields).forEach((subfield) => {
             if (!subfield.validate(subfield.value)) {
@@ -133,53 +113,77 @@ export default {
     },
   },
   methods: {
-    generatePasscode() {
-      const d = new Date();
-      const random = Date.now().toString();
-      this.passcode = `${this.selectedCompany.companyName.toString().charAt(0)}${random.substring(8, random.length)}-${d.getHours()}:${d.getUTCMinutes()}`; 
-    },
     setCompany(e) {
       this.selectedCompany = e;
     },
-    copyToClipboard() {
-      var copyText = this.$refs.passcode;
-      copyText.select();
-      copyText.setSelectionRange(0, 99999);
-      document.execCommand("copy");
-    },
     goLastStep() {
-      this.generatePasscode();
-      const allEnteredElements = this.$refs.form.querySelectorAll('.element');
-      allEnteredElements.forEach(element => {
+      if (!this.isFormValid) return;
+
+      /*
+       * The reason of usage `allEnteredFields`
+       *
+       * `formElements` variable stores every value & fields data at the same time
+       * At the post action, we need to get only filled field values as an object to post it to BE
+       * For this reason, we are not mapping `formElements` directly,
+       * Instead, we fill `allEnteredFields` with filled fields' value to have one seperate object
+       */
+      const allEnteredFields = this.$refs.form.querySelectorAll('.element');
+      allEnteredFields.forEach((element) => {
         this.enteredData[element.getAttribute('data-field')] = element.getAttribute('value');
       });
-      setTimeout(() => {
-        window.scrollTo(0, window.innerHeight);
-      }, 100);
+
+      setTimeout(() => window.scrollTo(0, window.innerHeight), 100);
     },
-    finishCaseCreation() {
+    resetForm() {
+      this.enteredData = [];
+      const formFieldsWithEmptyValue = mapValues(this.formFields, (field) => {
+        // Avoid resetting feedbackType
+        if (field.id === 'feedbackType') {
+          return {
+            ...field,
+          };
+        }
+
+        return {
+          ...field,
+          value: '',
+          subfields: {
+            ...mapValues(field.subfields, (subfield) => {
+              return {
+                ...subfield,
+                value: '',
+              };
+            }),
+          },
+        };
+      });
+      this.formFields = formFieldsWithEmptyValue;
+    },
+    onInputChange(input) {
+      if (input.id === 'feedbackType') {
+        this.resetForm();
+      }
+    },
+    finishCaseCreation(passcode) {
       this.loading = true;
       const payload = {
         ...this.enteredData,
-        passcode: this.passcode,
         company: this.selectedCompany,
         created: firebase.firestore.FieldValue.serverTimestamp(),
         status: true,
+        passcode,
       };
       API.cases.post(payload)
         .then((doc) => {
           localStorage.setItem('caseId', doc.id);
-          setTimeout(() => {
-            this.$router.push({
-              name: 'CaseDetail',
-            })
-          }, 400)
+          setTimeout(() => this.$router.push({ name: 'CaseDetail' }), 400);
           this.loading = false;
         })
         .catch((error) => {
           console.warn(error);
           this.loading = false;
-        });
+        })
+        .finally(() => this.resetForm());
     },
   },
 };
@@ -231,78 +235,4 @@ export default {
     }
   }
 }
-
-.report-page__warning {
-  width: 100%;
-  background: var(--dark-black);
-  flex-direction: row;
-  display: flex;
-  align-items: flex-start;
-  padding: 50px 20%;
-  padding-bottom: 200px;
-}
-
-
-.report-page__content {
-  text-align: left;
-  color: white;
-  padding-left: 50px;
-
-  strong {
-    font-family: inherit;
-    font-size: 36px;
-    line-height: 44px;
-    font-weight: 500;
-  }
-
-  p {
-    font-size: 16px;
-  }
-}
-
-.report-page__passcode {
-  background: white;
-  padding: 12px 0px 12px 20px;
-  color: black;
-  width: fit-content;
-  font-size: 35px;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  display: flex;
-
-  span {
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none; 
-    margin-right: 8px;
-  }
-
-  input {
-    padding: 5px 20px 5px 20px;
-    background: #ececec;
-    text-align: left;
-    color: red;
-    max-width: max-content;
-    margin: 0;
-    font-size: 30px;
-    max-width: 260px;
-  }
-
-  button, span {
-    font-size: 18px;
-    font-weight: 600;
-    outline: none;
-    opacity: .7;
-  }
-
-  button {
-    cursor: pointer;
-    background: transparent;
-    border: none;
-    padding: 0 15px;
-  }
-}
-
 </style>
